Refactor: Move LP code together.           Drop Python 2.7 support.
