+++ runestone/lp/__init__.py
@@ -1,10 +1,7 @@
-import sys
-from runestone.lp import inlinesyntaxhighlight
+from runestone.lp import inlinesyntaxhighlight, lp
+
 
 # Initialize both extensions.
 def setup(app):
     inlinesyntaxhighlight.setup(app)
-    # The LP extension requires Python 3.
-    if sys.version_info >= (3, 3):
-        from runestone.lp import lp
-        return lp.setup(app)
+    return lp.setup(app)
