+++ runestone/lp/lp.py
@@ -1,19 +1,5 @@
-# .. Copyright (C) 2017 Bryan A. Jones.
+# .. Copyright (C) 2019 Bryan A. Jones.
 #
-#    This file is part of E-Book Binder.
-#
-#    E-Book Binder is free software: you can redistribute it and/or modify it
-#    under the terms of the GNU General Public License as published by the Free
-#    Software Foundation, either version 3 of the License, or (at your option)
-#    any later version.
-#
-#    E-Book Binder is distributed in the hope that it will be useful, but WITHOUT
-#    ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
-#    FOR A PARTICULAR PURPOSE.  See the GNU General Public License for more
-#    details.
-#
-#    You should have received a copy of the GNU General Public License along
-#    with E-Book Binder.  If not, see <http://www.gnu.org/licenses/>.
 #
 # .. highlight:: python
 #
@@ -49,7 +35,7 @@ from CodeChat.CodeToRest import get_lexer
 # Local imports
 # -------------
 from ..common.runestonedirective import RunestoneIdDirective, RunestoneNode
-from ..server.lp_common_lib import STUDENT_SOURCE_PATH, code_here_comment, SPHINX_CONFIG_NAME
+from .lp_common_lib import STUDENT_SOURCE_PATH, code_here_comment, SPHINX_CONFIG_NAME
 from ..server.componentdb import addQuestionToDB, addHTMLToDB
 #
 # Directives
